# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Aiome Management Console API",
    "description": "Core API for the Autonomous AI Operating System",
    "license": {
      "name": "BUSL-1.1",
      "identifier": "BUSL-1.1"
    },
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/artifacts": {
      "get": {
        "tags": [
          "crate::routes::artifacts"
        ],
        "operationId": "list_artifacts_handler",
        "parameters": [
          {
            "name": "category",
            "in": "query",
            "description": "Filter by category",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "q",
            "in": "query",
            "description": "Semantic search query",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "description": "Limit results",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of artifacts",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {}
                }
              }
            }
          }
        }
      }
    },
...
    "/api/commune/status": {
      "get": {
        "tags": [
          "crate::routes::commune"
        ],
        "operationId": "commune_status",
        "responses": {
          "200": {
            "description": "Status retrieved",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        },
        "security": [
          {
            "api_key": []
          }
        ]
      }
    },
...
    "/api/synergy/test/failure": {
      "post": {
        "tags": [
          "crate::routes::karma"
        ],
        "operationId": "trigger_failure_demo",
        "responses": {
          "200": {
            "description": "Demo failure triggered",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        },
        "security": [
          {
            "api_key": []
          }
        ]
      }
    },
    "/api/synergy/test/federation": {
      "post": {
        "tags": [
          "crate::routes::karma"
        ],
        "operationId": "trigger_federation_demo",
        "responses": {
          "200": {
            "description": "Demo federation triggered",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        },
        "security": [
          {
            "api_key": []
          }
        ]
      }
    },
    "/api/synergy/test/security": {
      "post": {
        "tags": [
          "crate::routes::karma"
        ],
        "operationId": "trigger_security_demo",
        "responses": {
          "200": {
            "description": "Demo security triggered",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        },
        "security": [
          {
            "api_key": []
          }
        ]
      }
    },
...
    "/api/v1/audit/diagnostics": {
      "get": {
        "tags": [
          "crate::routes::audit"
        ],
        "operationId": "get_diagnoses",
        "responses": {
          "200": {
            "description": "Fetch agent diagnostics history",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/DiagnosisResponse"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          }
        },
        "security": [
          {
            "api_key": []
          }
        ]
      }
    },
    "/api/v1/audit/diagnostics/summary": {
      "get": {
        "tags": [
          "crate::routes::audit"
        ],
        "operationId": "get_diagnostics_summary",
        "responses": {
          "200": {
            "description": "Aggregated diagnostics summary",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DiagnosisSummaryResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          }
        },
        "security": [
          {
            "api_key": []
          }
        ]
      }
    },
...
    "/api/v1/audit/prompt-stats": {
      "get": {
        "tags": [
          "crate::routes::audit"
        ],
        "operationId": "get_audit_prompt_stats",
        "parameters": [
          {
            "name": "period",
            "in": "query",
            "description": "Statistics period (e.g., 7d)",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Fetch prompt evaluation stats",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PromptStatsResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          }
        },
        "security": [
          {
            "api_key": []
          }
        ]
      }
    },
    "/api/v1/audit/quarantine": {
      "get": {
        "tags": [
          "crate::routes::audit"
        ],
        "operationId": "get_quarantined_assets",
        "responses": {
          "200": {
            "description": "Fetch quarantined assets store",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/QuarantinedAsset"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          }
        },
        "security": [
          {
            "api_key": []
          }
        ]
      }
    },
    "/api/v1/audit/quarantine/{id}/release": {
      "post": {
        "tags": [
          "crate::routes::audit"
        ],
        "operationId": "release_quarantined_asset",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Asset ID to release",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Asset successfully released from quarantine"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          }
        },
        "security": [
          {
            "api_key": []
          }
        ]
      }
    },
...
    "/api/v1/demo/start": {
      "post": {
        "tags": [
          "crate::routes::demo"
        ],
        "summary": "デモの開始要請",
        "operationId": "start_demo",
        "responses": {
          "200": {
            "description": "Demo started",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          }
        },
        "security": [
          {
            "api_key": []
          }
        ]
      }
    },
...
    "/api/v1/settings": {
      "get": {
        "tags": [
          "crate::routes::settings"
        ],
        "operationId": "get_settings",
        "responses": {
          "200": {
            "description": "List all settings",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/SystemSetting"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          }
        },
        "security": [
          {
            "api_key": []
          }
        ]
      },
      "put": {
        "tags": [
          "crate::routes::settings"
        ],
        "operationId": "update_setting",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateSettingsRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Setting updated successfully"
          },
          "400": {
            "description": "Invalid request or unauthorized key"
          },
          "401": {
            "description": "Unauthorized"
          }
        },
        "security": [
          {
            "api_key": []
          }
        ]
      }
    },
    "/api/v1/settings/test": {
      "post": {
        "tags": [
          "crate::routes::settings"
        ],
        "operationId": "test_connection",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TestConnectionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Connection test completed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TestConnectionResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          }
        },
        "security": [
          {
            "api_key": []
          }
        ]
      }
    },
...
    "/api/v1/trends": {
      "get": {
        "tags": [
          "crate::routes::general"
        ],
        "operationId": "get_trends",
        "responses": {
          "200": {
            "description": "Fetch current trends from configured adapters (X, WebSearch, SERP)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TrendsResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          }
        },
        "security": [
          {
            "api_key": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "DiagnosisResponse": {
        "type": "object",
        "required": [
          "id",
          "job_id"
        ],
        "properties": {
          "failure_category": {
            "type": [
              "string",
              "null"
            ]
          },
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "job_id": {
            "type": "string"
          },
          "root_cause": {
            "type": [
              "string",
              "null"
            ]
          },
          "self_repair_hint": {
            "type": [
              "string",
              "null"
            ]
          },
          "timestamp": {
            "type": [
              "string",
              "null"
            ]
          }
        }
      },
      "DiagnosisSummaryResponse": {
        "type": "object",
        "required": [
          "total_diagnoses",
          "categories"
        ],
        "properties": {
          "categories": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CategoryCount"
            }
          },
          "total_diagnoses": {
            "type": "integer",
            "format": "int64"
          }
        }
      },
...
      "PromptStatsResponse": {
        "type": "object",
        "required": [
          "period",
          "providers"
        ],
        "properties": {
          "period": {
            "type": "string"
          },
          "providers": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ProviderEvalStat"
            }
          }
        }
      },
...
      "QuarantinedAsset": {
        "type": "object",
        "description": "検疫済みアセットのレコード構造体",
        "required": [
          "id",
          "asset_name",
          "image_hash",
          "reason",
          "status"
        ],
        "properties": {
          "asset_name": {
            "type": "string"
          },
          "id": {
            "type": "string"
          },
          "image_hash": {
            "type": "string"
          },
          "reason": {
            "type": "string"
          },
          "status": {
            "type": "string"
          },
          "uploaded_at": {
            "type": [
              "string",
              "null"
            ]
          }
        }
      },
...
    },
    "securitySchemes": {
      "api_key": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "API_SERVER_SECRET"
      }
    }
  }
}